move report function to a separate module
