+++ dff/utils/db_benchmark/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 # flake8: noqa: F401
-from dff.utils.db_benchmark.benchmark import report as context_storage_benchmark_report
+from dff.utils.db_benchmark.report import report
